# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job-App Backend API",
    "description": "REST API for job search platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/companies": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "List Companies",
        "description": "List all companies in the shared catalog.\n\nEndpoint: GET /companies\nAuth: Not required\n\nQuery Parameters:\n    - limit (int): Page size, 10-100, default 20\n    - offset (int): Page offset, default 0\n\nResponse (HTTP 200):\n    - companies: List of company objects\n    - total: Total number of companies\n    - hasMore: Boolean indicating if more results exist\n\nLogic:\n1. Scan Empresas table (all companies)\n2. Sort case-insensitively by nombre\n3. Return raw stored fields only (no derived/computed flags)\n4. Log retrieval operation\n\nCompany fields returned:\n    - companyId, nombre, careersUrl, plataforma\n    - lastScannedAt, lastScanStatus, lastVacancyCount, consecutiveFailures\n\nRequirements:\n- 6.1: Paginate listing\n- 6.2: Sort by nombre case-insensitively\n- 6.3: Return raw stored fields only\n- 6.4: Return {\"companies\", \"total\", \"hasMore\"}\n- 6.5: No auth required",
        "operationId": "list_companies_companies_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 10,
              "description": "Page size (10-100, default 20)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Page size (10-100, default 20)"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Page offset (default 0)",
              "default": 0,
              "title": "Offset"
            },
            "description": "Page offset (default 0)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompaniesListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "companies"
        ],
        "summary": "Add Company",
        "description": "Add a new company to the shared catalog.\n\nEndpoint: POST /companies\nAuth: Not required\n\nRequest Body:\n    - careersUrl (str): Company careers page URL (http/https)\n\nResponse (HTTP 201):\n    - companyId: SHA-256 hash of normalized URL\n    - nombre: Extracted company name\n    - plataforma: Detected platform (greenhouse, lever, html)\n    - createdAt: Timestamp\n\nError Responses:\n    - HTTP 400: Malformed URL or platform detection failed\n    - HTTP 409: Company already exists (duplicate URL)\n\nLogic:\n1. Validate careersUrl via validate_empresa_url (400 if invalid)\n2. Normalize URL and compute companyId (SHA-256 hash)\n3. Detect platform using ONLY hostname check (detect_platform_hostname_only)\n4. Check if companyId already exists (409 if found)\n5. Extract company nombre from URL domain (domain.com → \"domain\")\n6. Create Empresas entry with lastScannedAt=null, lastScanStatus=null, etc.\n7. Return 201 with companyId, nombre, plataforma, createdAt\n\nRequirements:\n- 7.1: Accept careersUrl\n- 7.2: Validate URL (400 on invalid)\n- 7.3: Normalize URL via normalize_url\n- 7.4: Compute companyId via compute_company_id (SHA-256)\n- 7.5: Detect platform via detect_platform_hostname_only (pure hostname check, no HTTP)\n- 7.6: Map platform detection failure to 400 platform_detection_failed\n- 7.7: Return 409 company_already_exists if hash exists\n- 7.8: Return 201 with companyId, nombre, plataforma, createdAt",
        "operationId": "add_company_companies_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCompanyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/companies/{company_id}": {
      "put": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Toggle Subscription",
        "description": "Activate or deactivate a company subscription for the user.\n\nEndpoint: PUT /me/companies/{companyId}\nAuth: Required (JWT)\n\nPath Parameters:\n    - companyId (str): Company ID (SHA-256 hash)\n\nRequest Body:\n    - activa (bool): True to activate, False to deactivate\n\nResponse (HTTP 200):\n    - companyId, activa, updatedAt\n\nError Responses:\n    - HTTP 404: Subscription not found for this (userId, companyId)\n    - HTTP 400: Company not found in Empresas table\n\nLogic:\n1. Validate subscription exists for (userId, companyId) (404 if missing)\n2. Validate companyId exists in Empresas (400 if missing)\n3. Set activa flag\n4. If reactivating (activa=true): Update addedAt to current timestamp\n5. If deactivating (activa=false): Keep addedAt unchanged\n6. Persist to Suscripciones\n7. Return 200 with companyId, activa, updatedAt\n\nRequirements:\n- 9.1: Accept PUT with activa flag\n- 9.2: Validate subscription exists (404 subscription_not_found)\n- 9.3: Validate company exists (400 company_not_found)\n- 9.4: Set activa flag\n- 9.5: Refresh addedAt on reactivation\n- 9.6: Persist atomically\n- 9.7: Return 200 with companyId, activa, updatedAt\n- 9.8: Log operation",
        "operationId": "toggle_subscription_me_companies__company_id__put",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddCompanyRequest": {
        "properties": {
          "careersUrl": {
            "type": "string",
            "title": "Careersurl",
            "description": "Company careers page URL (http/https)"
          }
        },
        "type": "object",
        "required": [
          "careersUrl"
        ],
        "title": "AddCompanyRequest",
        "description": "Request body for POST /companies.\n\nContains the careers URL for a company to be added to the shared catalog."
      },
...
      "CompaniesListResponse": {
        "properties": {
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyListItem"
            },
            "type": "array",
            "title": "Companies"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "hasMore": {
            "type": "boolean",
            "title": "Hasmore"
          }
        },
        "type": "object",
        "required": [
          "companies",
          "total",
          "hasMore"
        ],
        "title": "CompaniesListResponse",
        "description": "Response for GET /companies."
      },
      "CompanyCreateResponse": {
        "properties": {
          "companyId": {
            "type": "string",
            "title": "Companyid"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "plataforma": {
            "type": "string",
            "title": "Plataforma"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "companyId",
          "nombre",
          "plataforma",
          "createdAt"
        ],
        "title": "CompanyCreateResponse",
        "description": "Response for POST /companies (201 Created)."
      },
      "CompanyListItem": {
        "properties": {
          "companyId": {
            "type": "string",
            "title": "Companyid"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "careersUrl": {
            "type": "string",
            "title": "Careersurl"
          },
          "plataforma": {
            "type": "string",
            "title": "Plataforma"
          },
          "lastScannedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastscannedat"
          },
          "lastScanStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastscanstatus"
          },
          "lastVacancyCount": {
            "type": "integer",
            "title": "Lastvacancycount"
          },
          "consecutiveFailures": {
            "type": "integer",
            "title": "Consecutivefailures"
          }
        },
        "type": "object",
        "required": [
          "companyId",
          "nombre",
          "careersUrl",
          "plataforma",
          "lastVacancyCount",
          "consecutiveFailures"
        ],
        "title": "CompanyListItem",
        "description": "Company item in list response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubscriptionUpdateResponse": {
        "properties": {
          "companyId": {
            "type": "string",
            "title": "Companyid"
          },
          "activa": {
            "type": "boolean",
            "title": "Activa"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "companyId",
          "activa",
          "updatedAt"
        ],
        "title": "SubscriptionUpdateResponse",
        "description": "Response for PUT /me/companies/{companyId}."
      },
      "ToggleSubscriptionRequest": {
        "properties": {
          "activa": {
            "type": "boolean",
            "title": "Activa",
            "description": "Whether the subscription is active (true) or inactive (false)"
          }
        },
        "type": "object",
        "required": [
          "activa"
        ],
        "title": "ToggleSubscriptionRequest",
        "description": "Request body for PUT /me/companies/{companyId}.\n\nContains the subscription activation state."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}